backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.2.1
cryptography==46.0.3
cycler==0.12.1
dateparser==1.2.2
deprecation==2.1.0
durationpy==0.10
fastapi==0.104.1
fonttools==4.60.1
google-auth==2.41.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httmock==1.4.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
kiwisolver==1.4.9
kubernetes==29.13.0
matplotlib==3.10.7
multidict==6.7.0
numpy==2.3.4
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
postgrest==2.22.2
prometheus-api-client==0.5.5
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyGithub==2.1.1
PyJWT==2.10.1
PyNaCl==1.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.1
realtime==2.22.2
regex==2025.10.23
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
storage3==2.22.2
StrEnum==0.4.15
supabase==2.4.5
supabase-auth==2.22.2
supabase-functions==2.22.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn[standard]==0.24.0
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0deepdiff==8.6.1
